# src/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BadBeatMods API",
    "description": "This isn't really fully complete, but its better than absolutely nothing.\n\nThis API documentation is automatically generated and therefor may not be 100% accurate and may be missing a few fields. For example, request bodies are not fully fleshed out, and may not be accurate. Full documentation is still currently a work in progress.",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://bbm.saera.gay/api"
    }
  ],
...
  "paths": {
    "/beatmods/mod": {
      "get": {
        "tags": [
          "BeatMods"
        ],
        "summary": "Legacy BeatMods API endpoint.",
        "description": "Legacy BeatMods API endpoint. This is available for mod downloaders that have not been updated to use the new API.<br><br>This endpoint does not work the same way as the old BeatMods API, but it should be close enough to work with most mod downloaders.",
        "parameters": [
          {
            "name": "gameVersion",
            "description": "The game version as a string (ex. '1.29.1', '1.40.0').",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "The statuses to return. Available statuses are: 'approved' & 'all'",
            "format": "string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all mods."
          },
          "400": {
            "description": "Missing Game Version."
          }
        },
        "deprecated": true
      }
    },
    "/api/v1/mod": {
      "get": {
        "tags": [
          "BeatMods"
        ],
        "summary": "Legacy BeatMods API endpoint.",
        "description": "Legacy BeatMods API endpoint. This is available for mod downloaders that have not been updated to use the new API.<br><br>This endpoint does not work the same way as the old BeatMods API, but it should be close enough to work with most mod downloaders.",
        "parameters": [
          {
            "name": "gameVersion",
            "description": "The game version as a string (ex. '1.29.1', '1.40.0').",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "The statuses to return. Available statuses are: 'approved' & 'all'",
            "format": "string",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all mods."
          },
          "400": {
            "description": "Missing Game Version."
          }
        },
        "deprecated": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bbm_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}